coding to specification.
to allow the font rendering to satisfy my specification for
usual text rendering,

get_line_starting_from which uses a temporary string buffer and returns
the "slice" of string was created.

It's obviously not very good because of the temporary buffer, but it's
good enough for most reasons honestly... I guess I won't complain too
much.