Slope shit
Since I don't have a dedicated physics engine, this is kind of already a
PITA to do.

Although even with a dedicated physics engine, platformers still require
some special logic so thankfully I only need to get this to work and
we'll be good for a long time.

I'mma see if I can do a bit more work on getting this to be better, but
I might not be able to get slopes by tonight.