basic editor for the night.
really really basic, will work on more tomorrow probably. Might add
some graphical features.

I'm going to fix some of the bugs I noticed at the end of the week.

As well as start that opengl renderer near the end of this week. The
rest of the week is adding features and making tooling stuff though.