Changed loops for tilemap to be linear.
I wanted to see if this changed behavior so I could consider
a different storage form.

There "may" be slight advantages to organizing the storage like a 2D
Array but I'm not actually sure if it's worth making sure it is stored
in 2D array sorted order.

It would be easier to just random order it linearly. I can't think of
any times I directly ask for a specific tile anyhow...