Blackiron Console SFL added for quake console with no effort.
I previously wrote a quake console for my last game engine attempt,
and I found it was quite easy to break out into a SFL which I can
reuse over and over again.

It's written in C99, portable, and very easy to integrate and does
a good service for me... So why not?

Small visual bugs but it works just fine for what I'd need it for.

Unfortunately it's written in a very different style to what I used in
this project but not a big deal.