sparse particle storage.
It might be done slightly wrong since the last time I did any linked
list stuff was with the SFL console. Which was like half a year
ago. Seems to work fine, and allows me to allocate any amount of
particles up to the remaining arena memory.

Seems to not reuse chunks perfectly. So that'll need to be fixed but
at least I don't need to worry about my particle memory woes as much.

Performance seems to be very similar.

I did try to use contiguous ids (equivalent to
std::vector<Particle*>), but that shit killed my framerate really
bad. So this was the best option.