Better "continuous" collision
I shudder to call it continuous, however it's more continuous than
before.

I take additional substeps and "raycast" in a minkowski sort of way to
do the snapping. This now actually looks good.

Will simplify after fixing the jumping, which I'm going to start to
treat as an "impulse", IE: I apply it after we evaluate some physics?