Set the stage for versioned tilemaps,
mostly so I don't have to restart completely from scratch all the
time. The format is usually pretty stable with me only adding, and
probably almost never removing.

So adding some extra code shouldn't be that big of a deal for versions
that don't match.

Should not be a breaking change :)