Plan to add a nice death screen
When I comeback I'll implement a save game system (a primitive one,
but one nonetheless. Same as Soulslike metroidvanias though).

The other systems can't really be implemented without more game design
time, but these systems are pretty easy to envision and work out for
the game's polished status.

Thankfully this is still my allocated "buffer" time, so I can afford
to slack a bit and take it easy for the first week.