Remove death_animation procs
I do not have a good way to architect it like this, so we'll do
the simple thing, and just bake it into a drawing procedure which is
kind of exactly how these things should be drawn.

The player's death will still be heavily special cased, so yeah but
otherwise we do this.